nsal/signature.go: Move SignaturePolicy to internal
