+++ xal/internal/request.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/df-mc/go-xsapi/v2/xal"
 	"github.com/df-mc/go-xsapi/v2/xal/internal/timestamp"
-	"github.com/df-mc/go-xsapi/v2/xal/nsal"
 	"github.com/go-jose/go-jose/v4"
 )
 
@@ -45,7 +44,7 @@ func (r TokenRequest[P]) Do(ctx context.Context, config xal.Config, reqURL strin
 	req.Header.Set("User-Agent", config.UserAgent)
 	req.Header.Set("Content-Type", "application/json")
 	req.Header.Set("x-xbl-contract-version", "1")
-	if err := nsal.AuthPolicy.Sign(req, buf.Bytes(), proofKey, timestamp.Now()); err != nil {
+	if err := AuthPolicy.Sign(req, buf.Bytes(), proofKey, timestamp.Now()); err != nil {
 		return fmt.Errorf("sign request: %w", err)
 	}
